api/requirements.txt
fastapi>=0.110.0
python-multipart>=0.0.9
python-docx>=1.1.0
openpyxl>=3.1.2
lxml>=5.1.0
